lth_reproduction/requirements.txt
future==0.18.2
gin-config==0.3.0
numpy==1.18.3
Pillow==7.1.2
six==1.14.0
torch==1.4.0
torchvision==0.5.0
pytorch-ignite==0.3.0
tensorboardX==2.0


